fix(bar): prevent komorebi connection from staling
Sometimes the bar would randomly stop receiving notifications from
komorebi and would stop updating the `Komorebi` widget. This feels to me
that the reason is the same one that used to happen on the
`process_commands` from `komorebi` where the socket would get stuck
reading an empty connection. This commit adds a read timeout to the
socket to prevent that from happening and hopefully it should stop those
situations where the bar would stop receiving notifications.