fix(wm): properly store tile state when caching ws
Previously, when caching a workspace config for a monitor it would
simply store the `DefaultLayout` on `layout` even if the original
workspace config had the `layout` as `None` which makes komorebi create
a workspace with the `layout` as default `BSP` and the `tile` set at
`false`. However this wasn't being taken into consideration, so floating
workspaces would becoming tiling `BSP` workspaces after a monitor
disconnect and reconnect. This commit fixes this by turning the `layout`
to `None` when `tile` is `false`.