fix(wm): correctly use the display_index_preferences
This commit reworks the way the `postload` and the `reload` functions
apply the monitor configs to the monitors. Previously it was looping
through the monitor configs and applying them to the monitor with the
index corresponding to the config's index. However this isn't correct,
since the user might set the preferred indices for 3 monitors (like
monitor A, B and C), with the preferred index set to 0 for A, 1 for B
and 2 for C, but if only monitors A and C are connected then komorebi
would apply config 0 to A and config 1 to C, which is wrong it should be
2 for C.

This commit changes the way the configs are applied on those functions.
Now it loops through the existing monitors (already in order), then
checks if the monitor has a preferred config index, if it does it uses
that one, if it doesn't then it uses the first monitor config that isn't
a preferred index for some other monitor and that hasn't been used yet.

For the situation above it means that it would still apply config 2 to
monitor C. And in case there aren't any display_index_preferences set it
will still apply the configs in order.