fix(wm): restore orphaned containers
When a monitor was disconnected the containers from the removed monitor
were being moved to the primary monitor. However they weren't restored
so containers that were on an unfocused workspace of the removed monitor
would have been cloak and were getting added to the main monitor still
cloaked creating ghost tiles. This commit fixes that.