fix(bar): handle monitor disconnect/reconnect
This commit makes so a bar can be disabled when it's monitor is
disconnected and re-enabled when said monitor reconnects.
This commit also uses the new `monitor_usr_idx_map` to properly map the
monitor index given by the users on config to the actual monitor index
on the `WindowManager`, since in case some middle monitor gets
disconnected the index of the monitors to the "right" of that one will
be lowered by 1.
This should allow for cases with monitor A, B and C, if monitor B
disconnects, its bar will be disabled and monitor C will properly change
its monitor index internally to 1 so it still gets the infos from
monitor C (which now will have index 1 on the `WindowManager`).