fix(wm): cache monitor state instead of config
This commit changes the MONITOR_CACHE to hold an actual monitor with its
state instead of the config. This allows us to keep the state of a
disconnected monitor, so that when the monitor reconnects we still have
access to all its workspaces and windows as it was before.
While the monitor is disconnected, all the windows from that monitor are
not considered as handled by komorebi and so they're free to be handled
at any point, if for example the user uses `alt+tab` or presses the
taskbar icon of one of these windows that window will show up producing
it's `Show` event and will become handled by the currently focused
workspace of the focused monitor. When the disconnected monitor
reconnects it checks all windows and once it notices that this specific
window is now being shown on another monitor/workspace then it ignores
it.