fix(komorebi-bar-keyboard): fix retrival of current input method
Previously, current input method was retrieved via the Language::CurrentInputMethodLanguageTag()
call, which is inaccurate as it is retrieving the current input method of the bar itself, so it will
only be updated when the bar is in focus. The fix is getting the TID of the currently focused window
and read its input method.