feat(wm): toggle layer moves all windows
This commit makes it so when you toggle between workspace layers it
moves all windows of that layer to the top. So if you move to `Floating`
layer, then all floating windows are moved to the top, if you go back to
`Tiling` layer all containers are moved to the top.