fix(wm): stack on all directions even when on a stack
Previously the stacking logic would sometimes change the focused
container without actually changing focus to said container. This
resulted in the stack showing up with an unfocused border even though we
had focus on one windows of the stack (just not the right one).

Also before it wasn't possible to stack windows on some directions
when we were already on a stack.

This commit fixes both issues.