fix(bar): always add stroke on selected_frame
This commit fixes a breaking change on the selected_frame that was introduced by eframe version 0.31.

In this version, the stroke is drawn on the inside of a Frame instead it being drawn on the outside like before.
This now means that a stroke needs to be added on all the states of the Frame in order to avoid all the elements to be moving around on hover.