test(window_manger.rs): tests: WM Instance and Focus Workspace
Created a test that creates the WM instance and ensures the instance is running. The test creates a
custom socket and then cleans up the socket file after completion. Created a test that creates a WM
instance, monitor instance, and workpace. The tests checks to ensure that the expected workspace is
focused properly.