fix(bar): limit new state freq on time and date
This commit limits the number of times the time and date widgets get a new state.

Even though having a limit of 1 second on the ui repaint, when the bar is focused or hovered, this is ignored and state of these widgets were updated many more times a second.

The time widget looked less accurate and lagging (especially with multiple showing seconds) so the refresh interval is set to be 500 ms instead of 1 second.