test(window_manager.rs): remove Window From Container Test
Added a test that tests removing a window from a container. The test checks to ensure a new
container is created and that the removed window from the original container is added to the new
container. Also ensures that the old container has one less container after the function is called.