test(window_manager.rs): test Swapping Containers
Add a test to test swapping containers from different monitors. The test will create two monitors
and containers with different amount of windows. Test will verify that the swap is successful by
checking that the number of windows in the container matches the number of windows it was created
with in the previous monitor. The same will be done with the other container.