test(window_manager.rs): test Move Workspace To Monitor
Created a test to test moving a workspace from one monitor to another. The test will ensure that
after calling the function, the current monitor index is the one where the workspace was transfered
to. Test also will check the count to ensure the new workspace was added to the new monitor and that
the old monitor has one less.