feat(bar): auto select/hide widget based on value
This commit adds new settings to some widgets that allows to auto select/hide them based on their current values.

The cpu/memory/network/storage widgets get a setting that auto selects the widget if the current value/percentage is over a value.

The battery widget gets a setting that auto selects the widget if the current percentage is under a value.

The storage widget gets a setting that auto hides the disk widget if the percentage is under a value.

Also added 2 new settings (auto_select_fill and auto_select_text) to the theme, in order to select the fill and
text colors of an auto selected widget.

(Easter egg: the network icons change if the value is over the limit)

PR: #1353