fix(wm): fix bugs on workspace restore and remove workarounds
This commit removes some workarounds on the `update_focused_workspace`
function that were there to fix issues related to some bugs on the
`workspace.restore()` function. This commit fixes the bugs on the
`restore` function instead. The `update_focused_workspace` function
should be used only to update a workspace layout.