test(monitor_reconciliator/mod.rs): send Notification And Monitor Cache
Created a test to test sending a MonitorNotification using the send_notification function. The test
sends a MonitorNotification and tests that the notification was recieved using event_rx. Created a
test to test inserting a monitor into cache. The test creates a monitor and uses the
insert_in_monitor_cache function to insert the monitor into cache. The test will attempt to retrieve
the monitor from the cache and check to see if it matches the expected value.