test(monitor_reconciliator/mod.rs): tests for notification listener, channel capacity, monitor cache
Add test to test the channel capacity. The test will send 20 notifications and then another
notification. The test will check to ensure that all 20 messages match the notification we sent. The
test will also check to ensure the last notification sent wasn't received. Created a test for
testing the notification listener. The test will create a window manager and setup the notification
listener. The test will also check to see if we can send a notification and receive the
notification.