test(monitor_reconciliator/mod.rs): created Mock Monitor And attach_display_devices test
Created a Mock Monitor that can be used with testing functions that call the Win32 API. Created a
test to test the attach_display_devices function. The test will create a Mock Monitor and add the
monitor to a closure that simulates a display_provider. The test passes the display_provider into
the attach_display_devices function and checks to ensure we recieve the monitor we passed in.