test(workspace.rs): workspace tests
Added tests to test the focus_container_by_window and contains_managed_window functions. The
focus_container_by_window test, test the function by creating two containers with different distinct
windows. The test checks to see if we are focused on the expected container and window. The
contains_managed_window_test, tests by creating two containers and checks to ensure that when
calling the function returns the expected result when checking if a window is available in one of
the containers.